halfway refactoring service
